add view tour, diemthamquan, khachsan, nhahang, user crud

+++ DuLich/ViewModel/Models/DiemThamQuan.cs
@@ -6,7 +6,7 @@ namespace ViewModel.Models
     {
         public int ChuDichVu { get; set; }
         public string TenDiaDiem { get; set; }
-        public int Gia { get; set; }
+        public decimal Gia { get; set; }
         public string DiaDiem { get; set; }
         public string AnhDaiDien { get; set; }
         public DiemThamQuanCT DiemThamQuanCT { get; set; }
